Add fancy animations to sidebar toggle
